Create electricity_usage.m
Electricity data is stored in a file named electricity.mat. Load that MAT-file into MATLAB.
Residential data : res 
The commercial data : comm 
 Industrial data  : ind 
The usage data was collected annually between the years 1991 to 2013.
When looking at the output figure, it is clear that the industrial sector's electricity usage is fairly consistent and does not seem to fluctuate as much as the residential and commercial sectors.