Create audio_frequency.m
Audio signals are usually comprised of many different frequencies. For example, in music, the note 'middle C' has a fundamental frequency of 261.6 Hz, and most music consists of several notes (or frequencies) being played at the same time.

In this project, you will analyze the frequency content of an organ playing the C chord.

The C chord consists of the C (261.6 Hz), E (329.6 Hz), and G (392.0 Hz) notes. The highlighted points in this frequency plot correspond to each note.The C chord recording is stored in a file named Cchord.mat. This file contains two variables:
y: signal from recording
fs: sampling frequency
This task uses the numel function to return the number of elements in an array.
t now has the correct number of points, but it needs to represent the times when the audio signal was sampled. You can use the sampling frequency fs to convert the vector to time (in seconds).In the plot, notice that y is periodic, but it's not a simple sine wave. It's made up of multiple sine waves with different frequencies.
A Fourier transform will return information about the frequency content of the signal. The location of the dominant frequencies will show what notes are contained in the chord.
You can use the fft function to compute the discrete Fourier transform of a vector.
fft(y) . The output values from fft are complex numbers. You can use the abs function to get the magnitude.
In first two step you calculated the time vector t for the signal y. Similarly, you need to calculate the frequency vector f for your FFT vector yfft.The vector f now contains n points. To convert these points to frequencies, you can multiply the entire vector by the sampling frequency (fs) and divide it by the number of points (n).
f will contain frequences from 0 to fs. The dominant frequencies are located at the beginning of f. You can use the xlim function to zoom in on the area of interest.
xlim([xmin xmax])

The first three spikes in the OUTPUT graph are the notes comprising a middle C chord.
The other three spikes are, When a chord is played, the signal contains fundamental frequencies and the associated harmonics. In this case, the harmonics are another octave of the same chord.
Using the frequencies in the table below, you can see that the 6 spikes in the plot correspond to the fundamental frequencies and the first harmonics of a middle C chord.
Note	Frequency
C4                261.6
E4                329.6 
G4                392.0
C5                 525.3
E5                 659.3
G5                  784.0